organize the python types in mech config
