+++ python/musica/mechanism_configuration/__init__.py
@@ -1,18 +1,82 @@
-from .arrhenius import Arrhenius
-from .branched import Branched
-from .emission import Emission
-from .first_order_loss import FirstOrderLoss
+from .species import (
+    Species,
+    Phase,
+    PhaseSpecies,
+)
+from .reactions import (
+    Arrhenius,
+    Branched,
+    Emission,
+    FirstOrderLoss,
+    Photolysis,
+    ReactionComponent,
+    Reactions,
+    Surface,
+    TaylorSeries,
+    TernaryChemicalActivation,
+    Troe,
+    Tunneling,
+    UserDefined,
+)
+from .aerosol import (
+    Aerosol,
+    ArrheniusReferenceTemperature,
+    HenryLawConstant,
+    UniformSection,
+    SingleMomentMode,
+    TwoMomentMode,
+    DissolvedReaction,
+    DissolvedReversibleReaction,
+    HenryLawPhaseTransfer,
+    HenryLawEquilibrium,
+    DissolvedEquilibrium,
+    LinearConstraintTerm,
+    FixedConstant,
+    DiagnoseFromState,
+    LinearConstraint,
+)
+
 from .mechanism import Mechanism, Version, ReactionType
 from .parse import parse
-from .phase import Phase
-from .phase_species import PhaseSpecies
-from .photolysis import Photolysis
-from .reaction_component import ReactionComponent
-from .reactions import Reactions
-from .species import Species
-from .surface import Surface
-from .taylor_series import TaylorSeries
-from .ternary_chemical_activation import TernaryChemicalActivation
-from .troe import Troe
-from .tunneling import Tunneling
-from .user_defined import UserDefined
+
+__all__ = [
+    # species
+    "Species",
+    "Phase",
+    "PhaseSpecies",
+    # reactions
+    "Arrhenius",
+    "Branched",
+    "Emission",
+    "FirstOrderLoss",
+    "Photolysis",
+    "ReactionComponent",
+    "Reactions",
+    "Surface",
+    "TaylorSeries",
+    "TernaryChemicalActivation",
+    "Troe",
+    "Tunneling",
+    "UserDefined",
+    # aerosol
+    "Aerosol",
+    "ArrheniusReferenceTemperature",
+    "HenryLawConstant",
+    "UniformSection",
+    "SingleMomentMode",
+    "TwoMomentMode",
+    "DissolvedReaction",
+    "DissolvedReversibleReaction",
+    "HenryLawPhaseTransfer",
+    "HenryLawEquilibrium",
+    "DissolvedEquilibrium",
+    "LinearConstraintTerm",
+    "FixedConstant",
+    "DiagnoseFromState",
+    "LinearConstraint",
+    # top-level
+    "Mechanism",
+    "Version",
+    "ReactionType",
+    "parse",
+]
