+++ python/musica/mechanism_configuration/species/__init__.py
@@ -1,0 +1,9 @@
+from .species import Species
+from .phase import Phase
+from .phase_species import PhaseSpecies
+
+__all__ = [
+    "Species",
+    "Phase",
+    "PhaseSpecies",
+]
